license headers - format & clean up code

+++ CausticToolkit/src/main/java/com/teotigraphix/caustk/library/LibraryManager.java
@@ -1,3 +1,21 @@
+////////////////////////////////////////////////////////////////////////////////
+// Copyright 2013 Michael Schmalle - Teoti Graphix, LLC
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+// http://www.apache.org/licenses/LICENSE-2.0 
+// 
+// Unless required by applicable law or agreed to in writing, software 
+// distributed under the License is distributed on an "AS IS" BASIS, 
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and 
+// limitations under the License
+// 
+// Author: Michael Schmalle, Principal Architect
+// mschmalle at teotigraphix dot com
+////////////////////////////////////////////////////////////////////////////////
 
 package com.teotigraphix.caustk.library;
 
@@ -346,7 +364,7 @@ public class LibraryManager extends SubControllerBase implements ILibraryManager
             for (ToneDescriptor descriptor : set.getDescriptors()) {
                 int partIndex = descriptor.getIndex();
                 String name = descriptor.getName();
-                int patternBank = (int)(i / (64));
+                int patternBank = i / (64);
                 name = name + alpha[patternBank];
                 System.out.println("   " + name);
                 List<LibraryPhrase> list = library.findPhrasesByTagStartsWith(name);
